Move meshing into separate module
