+++ python/demos/demo_nitsche_rigid_surface_cuas.py
@@ -10,9 +10,8 @@ import numpy as np
 from mpi4py import MPI
 
 from dolfinx_contact.nitsche_rigid_surface_cuas import nitsche_rigid_surface_cuas
-from dolfinx_contact.create_contact_meshes import create_circle_plane_mesh, create_circle_circle_mesh,\
-    create_sphere_plane_mesh
-from dolfinx_contact.helpers import convert_mesh
+from dolfinx_contact.meshing import create_circle_plane_mesh, create_circle_circle_mesh,\
+    create_sphere_plane_mesh, convert_mesh
 
 if __name__ == "__main__":
     desc = "Nitsche's method with rigid surface using custom assemblers"
@@ -83,8 +82,8 @@ if __name__ == "__main__":
         if curved:
             fname = "two_disks"
             create_circle_circle_mesh(filename=f"{fname}.msh")
-            convert_mesh(fname, "triangle", prune_z=True)
-            convert_mesh(f"{fname}", "line", ext="facets", prune_z=True)
+            convert_mesh(fname, fname, "triangle", prune_z=True)
+            convert_mesh(f"{fname}", f"{fname}_facets" "line", prune_z=True)
 
             with XDMFFile(MPI.COMM_WORLD, f"{fname}.xdmf", "r") as xdmf:
                 mesh = xdmf.read_mesh(name="Grid")
@@ -125,8 +124,8 @@ if __name__ == "__main__":
         else:
             fname = "twomeshes"
             create_circle_plane_mesh(filename=f"{fname}.msh")
-            convert_mesh(fname, "triangle", prune_z=True)
-            convert_mesh(f"{fname}", "line", ext="facets", prune_z=True)
+            convert_mesh(fname, f"{fname}.xdmf", "triangle", prune_z=True)
+            convert_mesh(fname, f"{fname}_facets.xdmf", "line", prune_z=True)
 
             with XDMFFile(MPI.COMM_WORLD, f"{fname}.xdmf", "r") as xdmf:
                 mesh = xdmf.read_mesh(name="Grid")
