+++ python/dolfinx_contact/__init__.py
@@ -7,10 +7,9 @@
 # flake8: noqa
 
 
-from .create_mesh import create_disk_mesh, create_sphere_mesh
-from .helpers import lame_parameters, epsilon, sigma_func, convert_mesh
-
 from dolfinx_contact.cpp import Kernel, pack_circumradius_facet
 
-__all__ = ["create_disk_mesh", "create_sphere_mesh", "lame_parameters",
-           "epsilon", "sigma_func", "convert_mesh", "Kernel", "pack_circumradius_facet"]
+from .helpers import epsilon, lame_parameters, sigma_func
+
+__all__ = ["lame_parameters", "epsilon", "sigma_func", "Kernel",
+           "pack_circumradius_facet"]
