+++ python/dolfinx_contact/meshing/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (C) 2021  Jørgen S. Dokken
+#
+# SPDX-License-Identifier:    MIT
+
+from .contact_meshes import (create_box_mesh_2D,
+                             create_box_mesh_3D, create_circle_circle_mesh,
+                             create_circle_plane_mesh,
+                             create_sphere_plane_mesh)
+from .onesided_meshes import create_disk_mesh, create_sphere_mesh
+from .utils import convert_mesh
+
+__all__ = ["create_circle_plane_mesh", "create_circle_circle_mesh", "create_box_mesh_2D",
+           "create_box_mesh_3D", "create_sphere_plane_mesh", "convert_mesh", "create_disk_mesh",
+           "create_sphere_mesh"]
